LayoutTests/media/session/play-pause-media-events-in-content-sessions-expected.txt

'Content' sessions that are active should respond to Play/Pause media control events.


EVENT(canplaythrough)
EXPECTED (internals.mediaSessionCurrentState(session) == 'idle') OK
Idle Media Sessions should not respond to Play/Pause events.
EXPECTED (video.paused == 'true') OK
Sending Play/Pause media event.
RUN(internals.sendMediaControlEvent("play-pause"))
EXPECTED (video.paused == 'true') OK

Playing media.
EVENT(playing)
Media began playing.
EXPECTED (video.paused == 'false') OK
EXPECTED (internals.mediaSessionCurrentState(session) == 'active') OK
Sending Play/Pause media event.
RUN(internals.sendMediaControlEvent("play-pause"))
EXPECTED (video.paused == 'true') OK
END OF TEST
